+++ as2_map_server/plugins.xml
@@ -1,12 +1,7 @@
 <class_libraries>
-  <library path="scan2occ_grid">
-    <class type="scan2occ_grid::Plugin" base_class_type="as2_map_server_plugin_base::MapServerBase">
-      <description>Map server plugin for 2d mapping. From laser scan to occupancy grid.</description>
-    </class>
-  </library>
-  <library path="depth2occ_grid">
-    <class type="depth2occ_grid::Plugin" base_class_type="as2_map_server_plugin_base::MapServerBase">
-      <description>Map server plugin for 2d mapping. From depth image to occupancy grid.</description>
+  <library path="mapping_2d">
+    <class type="mapping_2d::Plugin" base_class_type="as2_map_server_plugin_base::MapServerBase">
+      <description>Map server plugin for 2d mapping.</description>
     </class>
   </library>
 </class_libraries>
