--- as2_map_server/plugins/scan2occ_grid/CMakeLists.txt
@@ -1,60 +1,0 @@
-cmake_minimum_required(VERSION 3.5)
-set(PLUGIN_NAME scan2occ_grid)
-
-# Default to C++17
-if(NOT CMAKE_CXX_STANDARD)
-  set(CMAKE_CXX_STANDARD 17)
-endif()
-
-# set Release as default
-if(NOT CMAKE_BUILD_TYPE)
-  set(CMAKE_BUILD_TYPE Release)
-endif()
-
-# find dependencies
-set(PLUGIN_DEPENDENCIES
-  ament_cmake
-  rclcpp
-  tf2
-  tf2_ros
-  tf2_geometry_msgs
-  geometry_msgs
-  nav_msgs
-  sensor_msgs
-  OpenCV
-)
-
-foreach(DEPENDENCY ${PLUGIN_DEPENDENCIES})
-  find_package(${DEPENDENCY} REQUIRED)
-endforeach()
-
-include_directories(
-  include
-  include/${PLUGIN_NAME}
-)
-
-add_library(${PLUGIN_NAME} SHARED src/${PLUGIN_NAME}.cpp)
-target_link_libraries(${PLUGIN_NAME} as2_map_server_base)
-ament_target_dependencies(${PLUGIN_NAME} ${PLUGIN_DEPENDENCIES})
-
-install(
-  DIRECTORY include/
-  DESTINATION include
-)
-
-ament_export_include_directories(include)
-ament_export_libraries(${PLUGIN_NAME})
-ament_export_targets(export_${PLUGIN_NAME})
-
-install(
-  TARGETS ${PLUGIN_NAME}
-  EXPORT export_${PLUGIN_NAME}
-  ARCHIVE DESTINATION lib
-  LIBRARY DESTINATION lib
-  RUNTIME DESTINATION bin
-)
-
-# PLUGIN TESTS
-# if(BUILD_TESTING)
-#   add_subdirectory(tests)
-# endif()
