+++ as2_msgs/msg/ControlMode.msg
@@ -16,6 +16,7 @@ int8 SPEED_IN_A_PLANE     =  4  # vx, vy , z refs
 int8 ATTITUDE             =  5  # roll, pitch, yaw  refs
 int8 ACRO                 =  6  # roll, pitch, yawrate  refs
 int8 TRAJECTORY           =  7  # x, y z, vx, vy, vz, ax, ay, az refs
+int8 ACEL                 =  8  # ax, ay, az refs
 
 int8 control_mode               # Control mode
