+++ as2_simulation_assets/as2_gazebo_assets/models/crazyflie/crazyflie.sdf
@@ -374,10 +374,5 @@ Copyright (c) 2022 Bitcraze
         </link>
     </model>
 
-    <joint name="imu_joint" type="fixed">
-      <parent>base_link</parent>
-      <child>imu</child>
-    </joint>
-
   </model>
 </sdf>
