--- as2_simulation_assets/as2_gazebo_assets/models/gate/gate.sdf
@@ -1,44 +1,0 @@
-<?xml version="1.0" ?>
-<sdf version="1.8">
-  <model name="gate">
-    <link name="gate_link">
-      <visual name="gate">
-        <pose>0 0 1.45 0 0 0</pose>
-        <geometry>
-          <mesh>
-            <uri>model://gate/meshes/gate.dae</uri>
-            <submesh>
-              <name>gate</name>
-            </submesh>
-          </mesh>
-        </geometry>
-        <material>
-          <diffuse>1.0 1.0 1.0 1.0</diffuse>
-          <specular>0.0 0.0 0.0 1.0</specular>
-          <pbr>
-            <metal>
-              <albedo_map>model://gate/meshes/gate_color.png</albedo_map>
-            </metal>
-          </pbr>
-        </material>
-      </visual>
-      <collision name="collision">
-        <pose>0 0 1.45 0 0 0</pose>
-        <geometry>
-          <mesh>
-            <uri>model://gate/meshes/gate.dae</uri>
-          </mesh>
-        </geometry>
-      </collision>
-    </link>
-
-    <static>1</static>
-    <plugin filename="gz-sim-pose-publisher-system" name="gz::sim::systems::PosePublisher">
-        <publish_model_pose>true</publish_model_pose>
-        <publish_link_pose>false</publish_link_pose>
-        <publish_nested_model_pose>true</publish_nested_model_pose>
-        <use_pose_vector_msg>true</use_pose_vector_msg>
-        <update_frequency>10</update_frequency>
-    </plugin>
-  </model>
-</sdf>
