--- as2_simulation_assets/as2_gazebo_assets/models/gate/model.config
@@ -1,19 +1,0 @@
-<?xml version="1.0" ?>
-<model>
-  <name>Gate</name>
-  <version>1.0</version>
-  <sdf sdf="1.8">gate.sdf</sdf>
-
-  <author>
-    <name>Pedro Arias Pérez</name>
-    <email>pedro.ariasp@upm.es</email>
-  </author>
-
-  <author>
-    <name>Rodrigo Da Silva</name>
-  </author>
-
-  <description>
-    CVAR gate model. Generated by blender SDF tools.
-  </description>
-</model>
