--- as2_simulation_assets/as2_gazebo_assets/models/px4vision/model.config
@@ -1,20 +1,0 @@
-<?xml version="1.0"?>
-<model>
-  <name>PX4 Vision</name>
-  <version>1.0</version>
-  <sdf version='1.9'>px4vision.sdf</sdf>
-
-  <author>
-   <name>Jaeyoung Lim</name>
-   <email>jaeyoung@auterion.com</email>
-  </author>
-
-  <author>
-   <name>Francisco J. Angutia</name>
-   <email>fj.anguita@alumnos.upm.es</email>
-  </author>
-
-  <description>
-    This is a model of the Holybro PX4 Vision adapted to the Aerostack2 framework.
-  </description>
-</model>
