+++ as2_simulation_assets/as2_gazebo_assets/src/as2_gazebo_assets/models/drone.py
@@ -63,7 +63,7 @@ class DroneTypeEnum(str, Enum):
     HEXROTOR = 'hexrotor_base'
     CRAZYFLIE = 'crazyflie'
     X500 = 'x500'
-    PX4 = 'px4vision'
+    A2RL = 'drone_a2rl'
 
 
 class Drone(Entity):
