+++ as2_user_interfaces/as2_alphanumeric_viewer/src/alphanumeric_viewer.cpp
@@ -1118,6 +1118,9 @@ void AlphanumericViewer::printControlModeInControl()
     case as2_msgs::msg::ControlMode::TRAJECTORY:
       printw("TRAJECTORY   ");
       break;
+    case as2_msgs::msg::ControlMode::ACEL:
+      printw("ACEL         ");
+      break;
     default:
       printw("UNKNOWN      ");
       break;
@@ -1189,6 +1192,9 @@ void AlphanumericViewer::printControlModeOutControl()
       break;
     case as2_msgs::msg::ControlMode::TRAJECTORY:
       printw("TRAJECTORY   ");
+      break;
+    case as2_msgs::msg::ControlMode::ACEL:
+      printw("ACEL         ");
       break;
     default:
       printw("UNKNOWN      ");
